Fix controllers and request
