+++ app/Http/Controllers/Api/CategoryController.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
 use App\Models\Category;
 use App\Http\Controllers\BaseController;
-use App\Http\Requests\Category\StoreRequest;
-use App\Http\Requests\Category\UpdateRequest;
+use App\Http\Requests\CategoryRequest;
 
 class CategoryController extends BaseController
 {
@@ -16,7 +15,7 @@ class CategoryController extends BaseController
         return $this->sendResponse($categories,"LIST OF CATEGORIES SUCCESSFULLY RETRIEVED",200);
     }
 
-    public function store(StoreRequest $request)
+    public function store(CategoryRequest $request)
     {
         $category = Category::create($request->all());
         return $this->sendResponse($category,"CREATED CATEGORY",201);
@@ -27,7 +26,7 @@ class CategoryController extends BaseController
         return $this->sendResponse($category,"RETRIEVED CATEGORY",200);
     }
 
-    public function update(UpdateRequest $request, Category $category)
+    public function update(CategoryRequest $request, Category $category)
     {
         $category->update($request->all());
         return $this->sendResponse($category,"UPDATED CATEGORY",201);
