+++ routes/api.php
@@ -17,20 +17,20 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::get('clients'            ,'ClientController@index');
-Route::get('clients/{client}'   ,'ClientController@show');
-Route::post('clients'           ,'ClientController@store');
-Route::put('clients/{client}'   ,'ClientController@update');
-Route::delete('clients/{client}','ClientController@destroy');
+Route::get('clients'            ,'Api\ClientController@index');
+Route::get('clients/{client}'   ,'Api\ClientController@show');
+Route::post('clients'           ,'Api\ClientController@store');
+Route::put('clients/{client}'   ,'Api\ClientController@update');
+Route::delete('clients/{client}','Api\ClientController@destroy');
 
-Route::get('categories'              ,'CategoryController@index');
-Route::get('categories/{category}'   ,'CategoryController@show');
-Route::post('categories'             ,'CategoryController@store');
-Route::put('categories/{category}'   ,'CategoryController@update');
-Route::delete('categories/{category}','CategoryController@destroy');
+Route::get('categories'              ,'Api\CategoryController@index');
+Route::get('categories/{category}'   ,'Api\CategoryController@show');
+Route::post('categories'             ,'Api\CategoryController@store');
+Route::put('categories/{category}'   ,'Api\CategoryController@update');
+Route::delete('categories/{category}','Api\CategoryController@destroy');
 
-Route::get('payment_modes'                 ,'PaymentModeController@index');
-Route::get('payment_modes/{paymentMode}'   ,'PaymentModeController@show');
-Route::post('payment_modes'                ,'PaymentModeController@store');
-Route::put('payment_modes/{paymentMode}'   ,'PaymentModeController@update');
-Route::delete('payment_modes/{paymentMode}','PaymentModeController@destroy');
+Route::get('payment_modes'                 ,'Api\PaymentModeController@index');
+Route::get('payment_modes/{paymentMode}'   ,'Api\PaymentModeController@show');
+Route::post('payment_modes'                ,'Api\PaymentModeController@store');
+Route::put('payment_modes/{paymentMode}'   ,'Api\PaymentModeController@update');
+Route::delete('payment_modes/{paymentMode}','Api\PaymentModeController@destroy');
